+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 ## 23.06.26
 
 added **frameTimer** + **Particle system** + **Update README.md** + **Improved work with Vec1** + **Fixed a bug with not displaying an object per frame**
+
+# 24.06.26
+
+updated **Simple particles**
+added **randomColor3** + **randomColor4**